Moved install helpers out of esig package main
